build: custom-hook useRecordinateQuickMenu
