+++ packages/web/src/custom-hooks/quickMenu/index.jsx
@@ -1,4 +1,5 @@
 import useQuickMenu from './useQuickMenu';
 import useQuickMenuListener from './useQuickMenuListener';
+import useRecordinateQuickMenu from './useRecordinateQuickMenu';
 
-export { useQuickMenu, useQuickMenuListener };
+export { useQuickMenu, useQuickMenuListener, useRecordinateQuickMenu };
